requirements_web.txt
flask>=3.0.0
flask-cors>=4.0.0
pillow>=10.4.0
requests>=2.31.0
pywebview>=4.4.1
